requirements.txt
imgurpython==1.1.6
python-telegram-bot==2.8.7
beautifulsoup4==4.4.1
